+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 - discount from checking delay the time already spent checking.
 - reduce network timeout to 10 seconds.
 - improve fetching toots logic, reduce number of API calls.
+- implement database migrations
+- allow to mute/unmute Home timeline
 
 ### Fixed